Release of 1.9.3.1 (Bugfix release)
Bugfixes
- Filesystem fixes*
-- Fixed a bug which copy or moving of a file caused the loss of the
directory id
-- Iterative algorithms where replaced with recursive ones
-- Fixed a possibility of a sql injection while creating a new directory
-- Fixed a bug - it was possible to create a new dir without a name

Changes
- Moved operations into own functions (Program.inc.php)

Feature additions
- Now it's possible to rename folders, too.
- Added password recovery function over email (email - adress must been
set!)
- Added several new configuration values
- Added translation functions
-- Added english and german translation
-- English is default translation

Use with caution
This build is a bugfix release (because the problems with the filesystem
operations).
That's the reason why this build number is 1.9.3.1 and not 1.9.4.

The rating of this build is still "unstable". Use with caution!

* The filesystem bugfixes are still experimental. It is a very big
chance of loosing your data given. Until the build is not marked as
stable,
please consier twice to use the unstable build to host files at the
moment.