Update of the installer
This new version of the installer contains a better management in case
of errors. Now you get a detailed description of the problem which
appeared while installation. Also a bug was fixed which avoided the
second try of an installation. Further a bug was fixed which caused a
plain text display of the root password.